[STF][REFACTOR] remove duplicate code, move/rename test and packages
All HTML widget had duplicate methods and variables which I moved to the
abstract HTMLSTFRemoteObject.
I renamed the remotebot.widget and remotebot.widget.impl package to
htmlbot.widget and htmlbot.widget.impl
Also I renamed the ComponentViewTest to HtmlBasicWidgetTest and move it
to the StfSelfTest package.
