+++ de.fu_berlin.inf.dpp.ui.frontend/html/src/constants.jsx
@@ -3,7 +3,7 @@
 export const views = {
   MAIN: 'main-page',
   ADD_CONTACT: 'add-contact',
-  START_SESSION_WIZARD: 'start-session-wizard',  
+  START_SESSION_WIZARD: 'start-session-wizard',
   CONFIGURATION_WIZARD: 'configuration-page',
-  COMPONENT_TEST: 'component-test'
+  BASIC_WIDGET_TEST: 'basic-widget-test'
 }
