+++ de.fu_berlin.inf.dpp.ui/test/framework/stf/src/de/fu_berlin/inf/dpp/stf/server/rmi/htmlbot/impl/HTMLBotImpl.java
@@ -9,8 +9,8 @@ import de.fu_berlin.inf.ag_se.browser.html.ISelector.CssClassSelector;
 import de.fu_berlin.inf.dpp.stf.server.HTMLSTFRemoteObject;
 import de.fu_berlin.inf.dpp.stf.server.bot.BotUtils;
 import de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.IHTMLBot;
-import de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.IRemoteHTMLView;
-import de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.impl.RemoteHTMLView;
+import de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.IRemoteHTMLView;
+import de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.impl.RemoteHTMLView;
 import de.fu_berlin.inf.dpp.ui.View;
 import de.fu_berlin.inf.dpp.ui.pages.MainPage;
