+++ de.fu_berlin.inf.dpp.ui/test/framework/stf/src/de/fu_berlin/inf/dpp/stf/server/rmi/htmlbot/widget/impl/RemoteHTMLInputField.java
@@ -1,24 +1,15 @@
-package de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.impl;
+package de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.impl;
 
 import java.rmi.RemoteException;
 
-import org.apache.log4j.Logger;
-
-import de.fu_berlin.inf.ag_se.browser.extensions.IJQueryBrowser;
-import de.fu_berlin.inf.ag_se.browser.html.ISelector;
-import de.fu_berlin.inf.ag_se.browser.html.ISelector.NameSelector;
 import de.fu_berlin.inf.dpp.stf.server.HTMLSTFRemoteObject;
-import de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.IRemoteHTMLInputField;
+import de.fu_berlin.inf.dpp.stf.server.bot.BotUtils;
+import de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.IRemoteHTMLInputField;
 
 public final class RemoteHTMLInputField extends HTMLSTFRemoteObject implements
     IRemoteHTMLInputField {
 
     private static final RemoteHTMLInputField INSTANCE = new RemoteHTMLInputField();
-    private static final Logger log = Logger
-        .getLogger(RemoteHTMLInputField.class);
-
-    private IJQueryBrowser browser;
-    private ISelector selector;
 
     public static RemoteHTMLInputField getInstance() {
         return INSTANCE;
@@ -26,14 +17,14 @@ public final class RemoteHTMLInputField extends HTMLSTFRemoteObject implements
 
     @Override
     public void enter(String text) throws RemoteException {
-        String name = getSelectorName();
+        String name = BotUtils.getSelectorName(selector);
         browser.val(selector, text);
         browser.run(String.format("view.setFieldValue('%s','%s')", name, text));
     }
 
     @Override
     public String getValue() throws RemoteException {
-        String name = getSelectorName();
+        String name = BotUtils.getSelectorName(selector);
         Object value = browser.syncRun(String.format(
             "return view.getFieldValue('%s')", name));
         return value != null ? value.toString() : null;
@@ -42,17 +33,5 @@ public final class RemoteHTMLInputField extends HTMLSTFRemoteObject implements
     @Override
     public void clear() throws RemoteException {
         enter("");
-    }
-
-    void setBrowser(IJQueryBrowser browser) {
-        this.browser = browser;
-    }
-
-    void setSelector(ISelector selector) {
-        this.selector = selector;
-    }
-
-    private String getSelectorName() {
-        return ((NameSelector) selector).getName();
     }
 }
