+++ de.fu_berlin.inf.dpp.ui/test/framework/stf/src/de/fu_berlin/inf/dpp/stf/server/rmi/htmlbot/widget/impl/RemoteHTMLRadioGroup.java
@@ -1,26 +1,16 @@
-package de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.impl;
+package de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.impl;
 
 import java.rmi.RemoteException;
 import java.util.List;
 
-import org.apache.log4j.Logger;
-
-import de.fu_berlin.inf.ag_se.browser.extensions.IJQueryBrowser;
-import de.fu_berlin.inf.ag_se.browser.html.ISelector;
-import de.fu_berlin.inf.ag_se.browser.html.ISelector.NameSelector;
 import de.fu_berlin.inf.dpp.stf.server.HTMLSTFRemoteObject;
 import de.fu_berlin.inf.dpp.stf.server.bot.BotUtils;
-import de.fu_berlin.inf.dpp.stf.server.rmi.remotebot.widget.IRemoteHTMLRadioGroup;
+import de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.IRemoteHTMLRadioGroup;
 
 public final class RemoteHTMLRadioGroup extends HTMLSTFRemoteObject implements
     IRemoteHTMLRadioGroup {
 
     private static final RemoteHTMLRadioGroup INSTANCE = new RemoteHTMLRadioGroup();
-    private static final Logger log = Logger
-        .getLogger(RemoteHTMLRadioGroup.class);
-
-    private IJQueryBrowser browser;
-    private ISelector selector;
 
     public static RemoteHTMLRadioGroup getInstance() {
         return INSTANCE;
@@ -28,7 +18,7 @@ public final class RemoteHTMLRadioGroup extends HTMLSTFRemoteObject implements
 
     @Override
     public String getSelected() throws RemoteException {
-        String name = getSelectorName();
+        String name = BotUtils.getSelectorName(selector);
         Object value = browser.syncRun(String.format(
             "return view.getFieldValue('%s')", name));
         return value != null ? value.toString() : null;
@@ -36,7 +26,7 @@ public final class RemoteHTMLRadioGroup extends HTMLSTFRemoteObject implements
 
     @Override
     public void select(String value) throws RemoteException {
-        String name = getSelectorName();
+        String name = BotUtils.getSelectorName(selector);
         browser.syncRun(String.format("view.setFieldValue('%s', '%s')", name,
             value));
     }
@@ -50,17 +40,4 @@ public final class RemoteHTMLRadioGroup extends HTMLSTFRemoteObject implements
     public List<String> values() throws RemoteException {
         return BotUtils.getListItemsValue(browser, selector);
     }
-
-    void setBrowser(IJQueryBrowser browser) {
-        this.browser = browser;
-    }
-
-    void setSelector(ISelector selector) {
-        this.selector = selector;
-    }
-
-    private String getSelectorName() {
-        return ((NameSelector) selector).getName();
-    }
-
 }
