+++ de.fu_berlin.inf.dpp/test/framework/stf/test/de/fu_berlin/inf/dpp/stf/test/stf/view/html/TestSuite.java
@@ -4,7 +4,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({ HtmlViewTest.class })
+@Suite.SuiteClasses({ HtmlViewTest.class, HtmlBasicWidgetTest.class })
 public class TestSuite {
     // the class remains completely empty,
     // being used only as a holder for the above annotations
