+++ de.fu_berlin.inf.dpp/test/stf/de/fu_berlin/inf/dpp/stf/test/html/TestSuite.java
@@ -4,8 +4,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({ AddContactTest.class, MainViewTest.class,
-    ComponentViewTest.class })
+@Suite.SuiteClasses({ AddContactTest.class, MainViewTest.class })
 public class TestSuite {
     // the class remains completely empty,
     // being used only as a holder for the above annotations
